+++ packages/autocomplete-client/src/search/index.ts
@@ -1,1 +1,1 @@
-export * from './fetchMeilisearchResults'
+export * from './fetchMeilisearchResults.js'