+++ packages/eslint-config-meilisearch/prettier.js
@@ -1,8 +1,0 @@
-module.exports = {
-  singleQuote: true,
-  arrowParens: 'always',
-  semi: false,
-  bracketSpacing: true,
-  trailingComma: 'es5',
-  printWidth: 80,
-}