+++ packages/instant-meilisearch/__tests__/placeholder-search.test.ts
@@ -1,5 +1,6 @@
-import { instantMeiliSearch } from '../src'
-import { dataset, Movies, meilisearchClient } from './assets/utils'
+import { describe, beforeAll, test, expect } from 'vitest'
+import { instantMeiliSearch } from '../src/index.js'
+import { dataset, type Movies, meilisearchClient } from './assets/utils.js'
 
 describe('Pagination browser test', () => {
   beforeAll(async () => {