+++ packages/instant-meilisearch/src/adapter/search-request-adapter/__tests__/geo-rules.test.ts
@@ -1,4 +1,5 @@
-import { adaptGeoSearch } from '../geo-rules-adapter'
+import { test, expect } from 'vitest'
+import { adaptGeoSearch } from '../geo-rules-adapter.js'
 
 test('Adapt instantsearch geo parameters to meilisearch filters without a boundingBox', () => {
   const filter = adaptGeoSearch({})