+++ packages/instant-meilisearch/src/adapter/search-request-adapter/index.ts
@@ -1,2 +1,2 @@
-export * from './search-resolver'
-export * from './search-params-adapter'
+export * from './search-resolver.js'
+export * from './search-params-adapter.js'