+++ packages/instant-meilisearch/src/adapter/search-response-adapter/__tests__/pagination-adapter.test.ts
@@ -1,5 +1,6 @@
-import { adaptPaginationParameters } from '../pagination-adapter'
-import { ceiledDivision } from './../__tests__/assets/number'
+import { describe, expect, it } from 'vitest'
+import { adaptPaginationParameters } from '../pagination-adapter.js'
+import { ceiledDivision } from './../__tests__/assets/number.js'
 
 const numberPagesTestParameters = [
   {