+++ packages/instant-meilisearch/src/contexts/sort-context.ts
@@ -1,15 +1,10 @@
 /**
  * Split sort string into an array.
  *
- * Example:
- * '_geoPoint(37.8153, -122.4784):asc,title:asc,description:desc'
+ * Example: '_geoPoint(37.8153, -122.4784):asc,title:asc,description:desc'
  *
- * becomes:
- * [
- * '_geoPoint(37.8153, -122.4784):asc',
- * 'title:asc',
- * 'description:desc',
- * ]
+ * Becomes: [ '_geoPoint(37.8153, -122.4784):asc', 'title:asc',
+ * 'description:desc', ]
  *
  * @param {string} sortStr
  * @returns {string[]}